remove redis dep from deps.ts and move to redisCache.ts
