+++ deps.ts
@@ -2,11 +2,6 @@ export { EventEmitter } from 'https://deno.land/x/event@0.2.1/mod.ts'
 export { unzlib } from 'https://denopkg.com/DjDeveloperr/denoflate@1.2/mod.ts'
 export { fetchAuto } from 'https://deno.land/x/fetchbase64@1.0.0/mod.ts'
 export { parse } from 'https://deno.land/x/mutil@0.1.2/mod.ts'
-export { connect } from 'https://deno.land/x/redis@v0.14.1/mod.ts'
-export type {
-  Redis,
-  RedisConnectOptions
-} from 'https://deno.land/x/redis@v0.14.1/mod.ts'
 export { walk } from 'https://deno.land/std@0.86.0/fs/walk.ts'
 export { join } from 'https://deno.land/std@0.86.0/path/mod.ts'
 export { Mixin } from 'https://esm.sh/ts-mixer@5.4.0'
